+++ lib/forest_liana/bootstraper.rb
@@ -1,5 +1,8 @@
+require_relative 'schema_file_updater'
+
 module ForestLiana
   class Bootstraper
+    SCHEMA_FILENAME = File.join(Dir.pwd, '.forestadmin-schema.json')
 
     def initialize
       @integration_stripe_valid = false
@@ -39,8 +42,24 @@ module ForestLiana
       require_lib_forest_liana
       format_and_validate_smart_actions
 
-      unless Rails.env.production?
-        create_apimap_json
+      if Rails.env.development?
+        @collections_sent = ForestLiana.apimap.as_json
+        @meta_sent = ForestLiana.meta
+        SchemaFileUpdater.new(SCHEMA_FILENAME, @collections_sent, @meta_sent).perform()
+      else
+        if File.exists?(SCHEMA_FILENAME)
+          begin
+            content = JSON.parse(File.read(SCHEMA_FILENAME))
+            @collections_sent = content['collections']
+            @meta_sent = content['meta']
+          rescue JSON::JSONError
+            FOREST_LOGGER.error "The content of .forestadmin-schema.json file is not a correct JSON."
+            FOREST_LOGGER.error "The schema cannot be synchronized with Forest Admin servers."
+          end
+        else
+          FOREST_LOGGER.error "The .forestadmin-schema.json file does not exists."
+          FOREST_LOGGER.error "The schema cannot be synchronized with Forest Admin servers."
+        end
       end
     end
 
@@ -152,7 +171,19 @@ module ForestLiana
         end
     end
 
+    def setup_forest_liana_meta
+      ForestLiana.meta = {
+        database_type: database_type,
+        framework_version: Gem.loaded_specs["rails"].version.version,
+        liana: 'forest-rails',
+        liana_version: liana_version,
+        orm_version: Gem.loaded_specs["activerecord"].version.version
+      }
+    end
+
     def create_apimap
+      setup_forest_liana_meta
+
       ForestLiana.models.map do |model|
         if analyze_model?(model)
           SchemaAdapter.new(model).perform
@@ -178,12 +209,6 @@ module ForestLiana
       end
     end
 
-    def create_apimap_json
-      File.open(File.join(Rails.root, 'forestadmin-schema.json'), 'w') do |f|
-        f.puts JSON.pretty_generate(ForestLiana.apimap.as_json)
-      end
-    end
-
     def require_lib_forest_liana
       path = Rails.root.join('lib', 'forest_liana', '**', '*.rb')
       Dir.glob(File.expand_path(path, __FILE__)).each do |file|
@@ -194,17 +219,6 @@ module ForestLiana
     def format_and_validate_smart_actions
       ForestLiana.apimap.each do |collection|
         collection.actions.each do |action|
-          if action.global
-            FOREST_LOGGER.warn "DEPRECATION WARNING: Smart Action \"global\" option is now " \
-              "deprecated. Please set \"type: 'global'\" instead of \"global: true\" for the " \
-              "\"#{action.name}\" Smart Action."
-          end
-
-          if action.type && !['bulk', 'global', 'single'].include?(action.type)
-            FOREST_LOGGER.warn "Please set a valid Smart Action type (\"bulk\", \"global\" or " \
-              "\"single\") for the \"#{action.name}\" Smart Action."
-          end
-
           if action.fields
             # NOTICE: Set a position to the Smart Actions fields.
             action.fields.each_with_index do |field, index|
@@ -216,18 +230,8 @@ module ForestLiana
     end
 
     def get_apimap_serialized
-      apimap = JSONAPI::Serializer.serialize(ForestLiana.apimap, {
-        is_collection: true,
-        include: ['actions', 'segments'],
-        meta: {
-          liana: 'forest-rails',
-          liana_version: liana_version,
-          framework_version: Gem.loaded_specs["rails"].version.version,
-          orm_version: Gem.loaded_specs["activerecord"].version.version,
-          database_type: database_type
-        }
-      })
-
+      serializer = ForestLiana::SchemaSerializer.new(@collections_sent, @meta_sent)
+      apimap = serializer.serialize
       ForestLiana::ApimapSorter.new(apimap).perform
     end
 
@@ -297,12 +301,10 @@ module ForestLiana
       model_name = ForestLiana.name_for(collection_name.constantize)
       # TODO: Remove once lianas prior to 2.0.0 are not supported anymore.
       model_name_old = ForestLiana.name_old_for(collection_name.constantize)
-      collection_display_name = collection_name.capitalize
 
       ForestLiana.apimap << ForestLiana::Model::Collection.new({
         name: "#{model_name}_intercom_conversations",
         name_old: "#{model_name_old}_intercom_conversations",
-        display_name: collection_display_name + ' Conversations',
         icon: 'intercom',
         integration: 'intercom',
         only_for_relationships: true,
@@ -320,30 +322,29 @@ module ForestLiana
       ForestLiana.apimap << ForestLiana::Model::Collection.new({
         name: "#{model_name}_intercom_attributes",
         name_old: "#{model_name_old}_intercom_attributes",
-        display_name: collection_display_name + ' Attributes',
         icon: 'intercom',
         integration: 'intercom',
         only_for_relationships: true,
         is_virtual: true,
         is_searchable: false,
         fields: [
-          { field: :created_at, type: 'Date', 'is-filterable': false },
-          { field: :updated_at, type: 'Date', 'is-filterable': false },
-          { field: :session_count, type: 'Number', 'is-filterable': false },
-          { field: :last_seen_ip, type: 'String', 'is-filterable': false },
-          { field: :signed_up_at, type: 'Date', 'is-filterable': false },
-          { field: :country, type: 'String', 'is-filterable': false },
-          { field: :city, type: 'String', 'is-filterable': false },
-          { field: :browser, type: 'String', 'is-filterable': false },
-          { field: :platform, type: 'String', 'is-filterable': false },
-          { field: :companies, type: 'String', 'is-filterable': false },
-          { field: :segments, type: 'String', 'is-filterable': false },
-          { field: :tags, type: 'String', 'is-filterable': false },
+          { field: :created_at, type: 'Date', is_filterable: false },
+          { field: :updated_at, type: 'Date', is_filterable: false },
+          { field: :session_count, type: 'Number', is_filterable: false },
+          { field: :last_seen_ip, type: 'String', is_filterable: false },
+          { field: :signed_up_at, type: 'Date', is_filterable: false },
+          { field: :country, type: 'String', is_filterable: false },
+          { field: :city, type: 'String', is_filterable: false },
+          { field: :browser, type: 'String', is_filterable: false },
+          { field: :platform, type: 'String', is_filterable: false },
+          { field: :companies, type: 'String', is_filterable: false },
+          { field: :segments, type: 'String', is_filterable: false },
+          { field: :tags, type: 'String', is_filterable: false },
           {
             field: :geoloc,
             type: 'String',
             widget: 'map',
-            'is-filterable': false
+            is_filterable: false
           }
         ]
       })
@@ -365,12 +366,10 @@ module ForestLiana
       model_name = ForestLiana.name_for(collection_name.constantize)
       # TODO: Remove once lianas prior to 2.0.0 are not supported anymore.
       model_name_old = ForestLiana.name_old_for(collection_name.constantize)
-      collection_display_name = model_name.capitalize
 
       ForestLiana.apimap << ForestLiana::Model::Collection.new({
         name: "#{model_name}_stripe_payments",
         name_old: "#{model_name_old}_stripe_payments",
-        display_name: collection_display_name + ' Payments',
         icon: 'stripe',
         integration: 'stripe',
         is_virtual: true,
@@ -378,23 +377,22 @@ module ForestLiana
         is_searchable: false,
         pagination_type: 'cursor',
         fields: [
-          { field: :id, type: 'String', 'is-filterable': false },
-          { field: :created, type: 'Date', 'is-filterable': false },
-          { field: :amount, type: 'Number', 'is-filterable': false },
-          { field: :status, type: 'String', 'is-filterable': false },
-          { field: :currency, type: 'String', 'is-filterable': false },
-          { field: :refunded, type: 'Boolean', 'is-filterable': false },
-          { field: :description, type: 'String', 'is-filterable': false },
+          { field: :id, type: 'String', is_filterable: false },
+          { field: :created, type: 'Date', is_filterable: false },
+          { field: :amount, type: 'Number', is_filterable: false },
+          { field: :status, type: 'String', is_filterable: false },
+          { field: :currency, type: 'String', is_filterable: false },
+          { field: :refunded, type: 'Boolean', is_filterable: false },
+          { field: :description, type: 'String', is_filterable: false },
           {
             field: :customer,
             type: 'String',
             reference: "#{model_name}.id",
-            'is-filterable': false
+            is_filterable: false
           }
         ],
         actions: [
           ForestLiana::Model::Action.new({
-            id: 'stripe.Refund',
             name: 'Refund',
             endpoint: '/forest/stripe_payments/refunds'
           })
@@ -404,7 +402,6 @@ module ForestLiana
       ForestLiana.apimap << ForestLiana::Model::Collection.new({
         name: "#{model_name}_stripe_invoices",
         name_old: "#{model_name_old}_stripe_invoices",
-        display_name: collection_display_name + ' Invoices',
         icon: 'stripe',
         integration: 'stripe',
         is_virtual: true,
@@ -412,26 +409,26 @@ module ForestLiana
         is_searchable: false,
         pagination_type: 'cursor',
         fields: [
-          { field: :id, type: 'String', 'is-filterable': false },
-          { field: :amount_due, type: 'Number', 'is-filterable': false },
-          { field: :attempt_count, type: 'Number', 'is-filterable': false },
-          { field: :attempted, type: 'Boolean', 'is-filterable': false },
-          { field: :closed, type: 'Boolean', 'is-filterable': false },
-          { field: :currency, type: 'String', 'is-filterable': false },
-          { field: :date, type: 'Date', 'is-filterable': false },
-          { field: :forgiven, type: 'Boolean', 'is-filterable': false },
-          { field: :period_start, type: 'Date', 'is-filterable': false },
-          { field: :period_end, type: 'Date', 'is-filterable': false },
-          { field: :subtotal, type: 'Number', 'is-filterable': false },
-          { field: :total, type: 'Number', 'is-filterable': false },
-          { field: :application_fee, type: 'Number', 'is-filterable': false },
-          { field: :tax, type: 'Number', 'is-filterable': false },
-          { field: :tax_percent, type: 'Number', 'is-filterable': false },
+          { field: :id, type: 'String', is_filterable: false },
+          { field: :amount_due, type: 'Number', is_filterable: false },
+          { field: :attempt_count, type: 'Number', is_filterable: false },
+          { field: :attempted, type: 'Boolean', is_filterable: false },
+          { field: :closed, type: 'Boolean', is_filterable: false },
+          { field: :currency, type: 'String', is_filterable: false },
+          { field: :date, type: 'Date', is_filterable: false },
+          { field: :forgiven, type: 'Boolean', is_filterable: false },
+          { field: :period_start, type: 'Date', is_filterable: false },
+          { field: :period_end, type: 'Date', is_filterable: false },
+          { field: :subtotal, type: 'Number', is_filterable: false },
+          { field: :total, type: 'Number', is_filterable: false },
+          { field: :application_fee, type: 'Number', is_filterable: false },
+          { field: :tax, type: 'Number', is_filterable: false },
+          { field: :tax_percent, type: 'Number', is_filterable: false },
           {
             field: :customer,
             type: 'String',
             reference: "#{model_name}.id",
-            'is-filterable': false
+            is_filterable: false
           }
         ]
       })
@@ -439,7 +436,6 @@ module ForestLiana
       ForestLiana.apimap << ForestLiana::Model::Collection.new({
         name: "#{model_name}_stripe_cards",
         name_old: "#{model_name_old}_stripe_cards",
-        display_name: collection_display_name + ' Cards',
         icon: 'stripe',
         integration: 'stripe',
         is_virtual: true,
@@ -448,26 +444,26 @@ module ForestLiana
         only_for_relationships: true,
         pagination_type: 'cursor',
         fields: [
-          { field: :id, type: 'String', 'is-filterable': false },
-          { field: :last4, type: 'String', 'is-filterable': false },
-          { field: :brand, type: 'String', 'is-filterable': false },
-          { field: :funding, type: 'String', 'is-filterable': false },
-          { field: :exp_month, type: 'Number', 'is-filterable': false },
-          { field: :exp_year, type: 'Number', 'is-filterable': false },
-          { field: :country, type: 'String', 'is-filterable': false },
-          { field: :name, type: 'String', 'is-filterable': false },
-          { field: :address_line1, type: 'String', 'is-filterable': false },
-          { field: :address_line2, type: 'String', 'is-filterable': false },
-          { field: :address_city, type: 'String', 'is-filterable': false },
-          { field: :address_state, type: 'String', 'is-filterable': false },
-          { field: :address_zip, type: 'String', 'is-filterable': false },
-          { field: :address_country, type: 'String', 'is-filterable': false },
-          { field: :cvc_check, type: 'String', 'is-filterable': false },
+          { field: :id, type: 'String', is_filterable: false },
+          { field: :last4, type: 'String', is_filterable: false },
+          { field: :brand, type: 'String', is_filterable: false },
+          { field: :funding, type: 'String', is_filterable: false },
+          { field: :exp_month, type: 'Number', is_filterable: false },
+          { field: :exp_year, type: 'Number', is_filterable: false },
+          { field: :country, type: 'String', is_filterable: false },
+          { field: :name, type: 'String', is_filterable: false },
+          { field: :address_line1, type: 'String', is_filterable: false },
+          { field: :address_line2, type: 'String', is_filterable: false },
+          { field: :address_city, type: 'String', is_filterable: false },
+          { field: :address_state, type: 'String', is_filterable: false },
+          { field: :address_zip, type: 'String', is_filterable: false },
+          { field: :address_country, type: 'String', is_filterable: false },
+          { field: :cvc_check, type: 'String', is_filterable: false },
           {
             field: :customer,
             type: 'String',
             reference: "#{model_name}.id",
-            'is-filterable': false
+            is_filterable: false
           }
         ]
       })
@@ -475,7 +471,6 @@ module ForestLiana
       ForestLiana.apimap << ForestLiana::Model::Collection.new({
         name: "#{model_name}_stripe_subscriptions",
         name_old: "#{model_name_old}_stripe_subscriptions",
-        display_name: collection_display_name + ' Subscriptions',
         icon: 'stripe',
         integration: 'stripe',
         is_virtual: true,
@@ -483,25 +478,25 @@ module ForestLiana
         is_searchable: false,
         pagination_type: 'cursor',
         fields: [
-          { field: :id, type: 'String', 'is-filterable': false },
-          { field: :cancel_at_period_end, type: 'Boolean', 'is-filterable': false },
-          { field: :canceled_at, type: 'Date', 'is-filterable': false },
-          { field: :created, type: 'Date', 'is-filterable': false },
-          { field: :current_period_end, type: 'Date', 'is-filterable': false },
-          { field: :current_period_start, type: 'Date', 'is-filterable': false },
-          { field: :ended_at, type: 'Date', 'is-filterable': false },
-          { field: :livemode, type: 'Boolean', 'is-filterable': false },
-          { field: :quantity, type: 'Number', 'is-filterable': false },
-          { field: :start, type: 'Date', 'is-filterable': false },
-          { field: :status, type: 'String', 'is-filterable': false },
-          { field: :tax_percent, type: 'Number', 'is-filterable': false },
-          { field: :trial_end, type: 'Date', 'is-filterable': false },
-          { field: :trial_start, type: 'Date', 'is-filterable': false },
+          { field: :id, type: 'String', is_filterable: false },
+          { field: :cancel_at_period_end, type: 'Boolean', is_filterable: false },
+          { field: :canceled_at, type: 'Date', is_filterable: false },
+          { field: :created, type: 'Date', is_filterable: false },
+          { field: :current_period_end, type: 'Date', is_filterable: false },
+          { field: :current_period_start, type: 'Date', is_filterable: false },
+          { field: :ended_at, type: 'Date', is_filterable: false },
+          { field: :livemode, type: 'Boolean', is_filterable: false },
+          { field: :quantity, type: 'Number', is_filterable: false },
+          { field: :start, type: 'Date', is_filterable: false },
+          { field: :status, type: 'String', is_filterable: false },
+          { field: :tax_percent, type: 'Number', is_filterable: false },
+          { field: :trial_end, type: 'Date', is_filterable: false },
+          { field: :trial_start, type: 'Date', is_filterable: false },
           {
             field: :customer,
             type: 'String',
             reference: "#{model_name}.id",
-            'is-filterable': false
+            is_filterable: false
           }
         ]
       })
@@ -509,7 +504,6 @@ module ForestLiana
       ForestLiana.apimap << ForestLiana::Model::Collection.new({
         name: "#{model_name}_stripe_bank_accounts",
         name_old: "#{model_name_old}_stripe_bank_accounts",
-        display_name: collection_display_name + ' Bank Accounts',
         icon: 'stripe',
         integration: 'stripe',
         is_virtual: true,
@@ -518,23 +512,23 @@ module ForestLiana
         only_for_relationships: true,
         pagination_type: 'cursor',
         fields: [
-          { field: :id, type: 'String', 'is-filterable': false },
-          { field: :account, type: 'String', 'is-filterable': false },
-          { field: :account_holder_name, type: 'String', 'is-filterable': false },
-          { field: :account_holder_type, type: 'String', 'is-filterable': false },
-          { field: :bank_name, type: 'String', 'is-filterable': false },
-          { field: :country, type: 'String', 'is-filterable': false },
-          { field: :currency, type: 'String', 'is-filterable': false },
-          { field: :default_for_currency, type: 'Boolean', 'is-filterable': false },
-          { field: :fingerprint, type: 'String', 'is-filterable': false },
-          { field: :last4, type: 'String', 'is-filterable': false },
-          { field: :rooting_number, type: 'String', 'is-filterable': false },
-          { field: :status, type: 'String', 'is-filterable': false },
+          { field: :id, type: 'String', is_filterable: false },
+          { field: :account, type: 'String', is_filterable: false },
+          { field: :account_holder_name, type: 'String', is_filterable: false },
+          { field: :account_holder_type, type: 'String', is_filterable: false },
+          { field: :bank_name, type: 'String', is_filterable: false },
+          { field: :country, type: 'String', is_filterable: false },
+          { field: :currency, type: 'String', is_filterable: false },
+          { field: :default_for_currency, type: 'Boolean', is_filterable: false },
+          { field: :fingerprint, type: 'String', is_filterable: false },
+          { field: :last4, type: 'String', is_filterable: false },
+          { field: :rooting_number, type: 'String', is_filterable: false },
+          { field: :status, type: 'String', is_filterable: false },
           {
             field: :customer,
             type: 'String',
             reference: "#{model_name}.id",
-            'is-filterable': false
+            is_filterable: false
           }
         ]
       })
@@ -572,9 +566,8 @@ module ForestLiana
     def setup_mixpanel_integration(collection_name_and_field)
       collection_name = collection_name_and_field.split('.')[0]
       model_name = ForestLiana.name_for(collection_name.constantize)
-      collection_display_name = model_name.capitalize
 
-      field_attributes = { 'is-filterable': false , 'is-virtual': true, 'is-sortable': false }
+      field_attributes = { is_filterable: false , is_virtual: true, is_sortable: false }
 
       fields = [
         { field: :id, type: 'String' },
@@ -601,7 +594,6 @@ module ForestLiana
 
       ForestLiana.apimap << ForestLiana::Model::Collection.new({
         name: "#{model_name}_mixpanel_events",
-        display_name: "#{collection_display_name} Events",
         icon: 'mixpanel',
         integration: 'mixpanel',
         is_virtual: true,