+++ lib/forest_liana/collection.rb
@@ -24,13 +24,11 @@ module ForestLiana::Collection
     end
 
     def action(name, opts = {})
-      opts[:id] = "#{self.collection_name.to_s}.#{name}"
       opts[:name] = name
       model.actions << ForestLiana::Model::Action.new(opts)
     end
 
     def segment(name, opts = {}, &block)
-      opts[:id] = "#{self.collection_name.to_s}.#{name}"
       opts[:name] = name
       model.segments << ForestLiana::Model::Segment.new(opts, &block)
     end
@@ -40,15 +38,46 @@ module ForestLiana::Collection
     end
 
     def field(name, opts, &block)
-      opts[:read_only] = true unless opts.has_key?(:read_only)
-      opts[:read_only] = false if opts.has_key?(:set)
+      # TODO: Handle empty name
+
+      if opts.key?(:isRequired)
+        FOREST_LOGGER.warn "DEPRECATION WARNING: isRequired on field \"#{name}\" is deprecated. Please use is_required."
+        opts[:is_required] = !!opts[:isRequired]
+        opts.delete(:isRequired)
+      end
+      if opts.key?(:isReadOnly)
+        FOREST_LOGGER.warn "DEPRECATION WARNING: isReadOnly on field \"#{name}\" is deprecated. Please use is_read_only."
+        opts[:is_read_only] = !!opts[:isReadOnly]
+        opts.delete(:isReadOnly)
+      end
+      if opts.key?(:isFilterable)
+        FOREST_LOGGER.warn "DEPRECATION WARNING: isFilterable on field \"#{name}\" is deprecated. Please use is_filterable."
+        opts[:is_filterable] = !!opts[:isFilterable]
+        opts.delete(:isFilterable)
+      end
+      if opts.key?(:isSortable)
+        FOREST_LOGGER.warn "DEPRECATION WARNING: isSortable on field \"#{name}\" is deprecated. Please use is_sortable."
+        opts[:is_sortable] = !!opts[:isSortable]
+        opts.delete(:isSortable)
+      end
+
+      opts[:is_read_only] = true unless opts.has_key?(:is_read_only)
+      opts[:is_read_only] = false if opts.has_key?(:set)
+      opts[:is_required] = false unless opts.has_key?(:is_required)
+      opts[:type] = "String" unless opts.has_key?(:type)
+      opts[:default_value] = nil unless opts.has_key?(:default_value)
+      opts[:integration] = nil unless opts.has_key?(:integration)
+      opts[:reference] = nil unless opts.has_key?(:reference)
+      opts[:inverse_of] = nil unless opts.has_key?(:inverse_of)
+      opts[:relationships] = nil unless opts.has_key?(:relationships)
+      opts[:widget] = nil unless opts.has_key?(:widget)
+      opts[:validations] = [] unless opts.has_key?(:validations)
 
       model.fields << opts.merge({
         field: name,
-        :'is-read-only' => opts[:read_only],
-        :'is-filterable' => !!opts[:is_filterable],
-        :'is-sortable' => !!opts[:is_sortable],
-        :'is-virtual' => true
+        is_filterable: !!opts[:is_filterable],
+        is_sortable: !!opts[:is_sortable],
+        is_virtual: true
       })
 
       define_method(name) { self.data[name] } if smart_collection?
@@ -80,8 +109,8 @@ module ForestLiana::Collection
     def has_many(name, opts, &block)
       model.fields << opts.merge({
         field: name,
-        :'is-virtual' => true,
-        :'is-searchable' => false,
+        is_virtual: true,
+        is_searchable: false,
         type: ['String']
       })
 
@@ -98,8 +127,8 @@ module ForestLiana::Collection
     def belongs_to(name, opts, &block)
       model.fields << opts.merge({
         field: name,
-        :'is-virtual' => true,
-        :'is-searchable' => false,
+        is_virtual: true,
+        is_searchable: false,
         type: 'String'
       })